backend/requirements.txt
pydantic
uvicorn
pymongo
python-dotenv
fastapi
passlib
bcrypt
python-jose
redis
email_validator
faker